Organize lib components by cohesion
