+++ src/lib/components/inspector/Inspector.svelte
