+++ src/lib/components/inspector/InspectorOverview.svelte
