+++ src/lib/components/statistics/StatisticsBanner.svelte
