+++ src/lib/components/student/StudentCard.svelte
